@@ -1,58 +1,61 @@
 {
   "name": "VXFrame",
   "version": "1.6.3",
-  "buildDate": "2022-07-02T10:01:00-07:00",
+  "buildDate": "2022-07-12T12:00:00-07:00",
   "description": "VXFrame",
   "author": "SOTA",
   "license": "MIT",
   "repository": "https://github.com/sotarules/vxframe",
   "dependencies": {
-    "@babel/runtime": "^7.11.2",
-    "ace-builds": "^1.4.12",
-    "async": "^3.2.0",
-    "aws-sdk": "^2.1048.0",
+    "@babel/runtime": "^7.18.6",
+    "ace-builds": "^1.7.1",
+    "async": "^3.2.4",
+    "aws-sdk": "^2.1172.0",
     "bcrypt": "^5.0.1",
-    "core-js": "^3.6.5",
-    "cropperjs": "^1.5.11",
+    "core-js": "^3.23.4",
+    "cropperjs": "^1.5.12",
     "domready": "^1.0.8",
     "fast-deep-equal": "^3.1.3",
-    "fast-safe-stringify": "^2.0.7",
-    "fibers": "^5.0.0",
+    "fast-safe-stringify": "^2.1.1",
+    "fibers": "^5.0.1",
     "filereader": "^0.10.3",
-    "html-react-parser": "^0.14.0",
-    "ladda": "^2.0.1",
+    "html-react-parser": "^3.0.1",
+    "ladda": "^2.0.3",
     "lodash": "^4.17.21",
-    "meteor-node-stubs": "^1.0.3",
-    "node-sass": "^4.14.1",
+    "meteor-node-stubs": "^1.2.3",
+    "mini-create-react-context": "^0.4.1",
+    "node-sass": "^7.0.1",
     "otplib": "^12.0.1",
-    "prop-types": "^15.7.2",
-    "qrcode.react": "^1.0.1",
-    "react": "^16.13.1",
-    "react-ace": "^9.2.1",
-    "react-contexify": "^4.1.1",
-    "react-dom": "^16.13.1",
+    "prop-types": "^15.8.1",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
+    "react-ace": "^10.1.0",
+    "react-contexify": "^5.0.0",
+    "react-dom": "^18.2.0",
     "react-fast-compare": "^3.2.0",
     "react-fastclick": "^3.0.2",
     "react-helmet": "^6.1.0",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-transition-group": "^4.4.1",
-    "readable-stream": "^3.6.0",
-    "redux": "^4.0.5",
+    "react-redux": "^8.0.2",
+    "react-router-dom": "^5.3.3",
+    "react-side-effect": "^2.1.2",
+    "readable-stream": "^4.1.0",
+    "redux": "^4.2.0",
     "redux-actions": "^2.6.5",
     "redux-persist": "^6.0.0",
-    "spin.js": "^4.1.0"
+    "simpl-schema": "^1.12.2",
+    "spin.js": "^4.1.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.18.6",
+    "@babel/eslint-parser": "^7.18.2",
     "@meteorjs/eslint-config-meteor": "^1.0.5",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^5.16.0",
-    "eslint-config-airbnb": "^18.2.0",
+    "eslint": "^8.19.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-import-resolver-meteor": "^0.4.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.3.1",
-    "eslint-plugin-meteor": "^7.2.0",
-    "eslint-plugin-react": "^7.21.2",
-    "eslint-plugin-react-hooks": "^4.1.2"
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.0",
+    "eslint-plugin-meteor": "^7.3.0",
+    "eslint-plugin-react": "^7.30.1",
+    "eslint-plugin-react-hooks": "^4.6.0"
   }
 }
