@@ -1,7 +1,7 @@
 import { Component } from "react"
 import PropTypes from "prop-types"
 import { Switch } from "react-router-dom"
-import Transition from "react-transition-group/Transition"
+import Transition from "/imports/vx/client/Transition"
 
 export default class RouteSlidePanel extends Component {
 
@@ -21,10 +21,11 @@ export default class RouteSlidePanel extends Component {
                 timeout={VXApp.animationTimeout(this.props.getAnimation())}>
                 {(status) => {
                     OLog.debug(`RouteSlidePanel.jsx render id=${this.props.id} key=${this.props.location.key} pathname=${this.props.location.pathname} ` +
-                        `in=${this.props.in} status=${status} animation=${this.props.getAnimation()}`)
+                        `in=${this.props.in} status=${status} animation=${this.props.getAnimation()} ` +
+                        `classes=${this.animationClasses(status)}`)
                     return (
                         <div id={this.props.id}
-                            className={VXApp.makeAnimationClasses(status, this.props.className, this.props.getAnimation())}>
+                            className={this.animationClasses(status)}>
                             <Switch location={this.props.location}>
                                 {this.props.children}
                             </Switch>
@@ -33,5 +34,9 @@ export default class RouteSlidePanel extends Component {
                 }}
             </Transition>
         )
+    }
+
+    animationClasses(status) {
+        return VXApp.makeAnimationClasses(status, this.props.className, this.props.getAnimation())
     }
 }
